requirements.txt
pip==24.3.1 
pypdf
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3

pyarrow==18.0.0
pyOpenSSL==23.2.0

python-dotenv==1.0.1
requests==2.32.3
virtualenv==20.25.0
zipp==3.20.2

pinecone
pinecone-client
pinecone-plugin-inference

langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4
langsmith==0.2.6
google-generativeai

